geotiff tests: consolidate GPU reader cluster (Sub-PR A of #2438)
Fold 10 GPU read-side regression files into a new
``xrspatial/geotiff/tests/gpu/test_reader.py``. Sections are grouped
by the upstream issue number they pin:

- #1605 -- window / band kwarg forwarding on GPU read
- stripped multiband (#1582) -- 3-D (y, x, band) reads via stripped fallback
- #1753 -- stripped no-georef windowed coord parity
- #1732 -- stripped fallback forwards max_pixels/window/band
- #1542 -- GPU read nodata promotion + attrs['nodata']
- #2097 -- write_geotiff_gpu rejects MinIsWhite (band-first guard)
- #1876 -- chunks= is a real out-of-core dask graph
- #2324 -- sidecar overview-inheritance georef parity
- #2161 -- HTTP / fsspec URLs on the eager GPU path
- #1909 -- GDS chunked path casts each chunk to declared dtype

Tests-only. Each file's local ``_gpu_available`` / ``_gpu_only`` helper
is replaced by the shared ``requires_gpu`` marker from
``_helpers/markers.py`` (aliased ``_gpu_only`` for ergonomic
decoration). 66 tests collected (matches the baseline sum of the 10
deleted files).

Cross-cutting touches:
- ``docs/source/reference/release_gate_geotiff.rst`` -- GPU nodata
  row now cites ``gpu/test_reader.py``.
- ``docs/source/reference/geotiff.rst`` -- prose pointer updated.
- ``read/test_nodata.py`` -- docstring pointer updated.

``CLUSTER_AUDIT_GPU_A.md`` maps every old ``file::test`` to its new
``test_id`` and is deleted on a final commit on this branch before
merge.
